# openapi.json
{"openapi":"3.1.0","info":{"title":"Bitcoin Address Generation API","description":"API for generating Bitcoin mnemonic seeds and various address types (BIP32, BIP44, BIP49, BIP84, BIP86, BIP85) including BIP141-compatible addresses.","version":"1.0.0"},"servers":[{"url":"http://127.0.0.1:8000","description":"Development server"},{"url":"https://btc-tx-gw.vercel.app","description":"Production environment"}],"paths":{"/":{"get":{"summary":"Read Root","description":"Root endpoint returning a simple greeting.","operationId":"read_root__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/generate-mnemonic":{"get":{"summary":"Generate Mnemonic","description":"Generate a new BIP39 mnemonic and seed.","operationId":"generate_mnemonic_generate_mnemonic_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MnemonicResponse"}}}}}}},"/generate-brain-wallet":{"post":{"summary":"Generate Brain Wallet Endpoint","description":"Generate a brain wallet from a passphrase.","operationId":"generate_brain_wallet_endpoint_generate_brain_wallet_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrainWalletRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BrainWalletResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generate-bip32-addresses":{"post":{"summary":"Generate Bip32 Addresses","description":"Generate BIP32 addresses from a mnemonic.","operationId":"generate_bip32_addresses_generate_bip32_addresses_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generate-bip44-addresses":{"post":{"summary":"Generate Addresses","description":"Generate BIP44 addresses from a mnemonic.","operationId":"generate_addresses_generate_bip44_addresses_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generate-bip49-addresses":{"post":{"summary":"Generate Bip49 Addresses","description":"Generate BIP49 (Wrapped SegWit P2SH-P2WPKH) addresses from a mnemonic.","operationId":"generate_bip49_addresses_generate_bip49_addresses_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generate-bip84-addresses":{"post":{"summary":"Generate Bip84 Addresses","description":"Generate BIP84 (Native SegWit P2WPKH) addresses from a mnemonic.","operationId":"generate_bip84_addresses_generate_bip84_addresses_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generate-bip86-addresses":{"post":{"summary":"Generate Bip86 Addresses","description":"Generate BIP86 (Taproot) addresses from a mnemonic.","operationId":"generate_bip86_addresses_generate_bip86_addresses_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generate-bip141-wrapped-segwit-via-bip49":{"post":{"summary":"Generate Bip141 Wrapped Segwit Via Bip49","description":"Generate BIP141-compatible Wrapped SegWit (P2SH-P2WPKH) addresses using BIP49.","operationId":"generate_bip141_wrapped_segwit_via_bip49_generate_bip141_wrapped_segwit_via_bip49_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generate-bip141-native-segwit-via-bip84":{"post":{"summary":"Generate Bip141 Native Segwit Via Bip84","description":"Generate BIP141-compatible Native SegWit (P2WPKH) addresses using BIP84.","operationId":"generate_bip141_native_segwit_via_bip84_generate_bip141_native_segwit_via_bip84_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/AddressListResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/generate-bip85-child-mnemonic":{"post":{"summary":"Generate Bip85 Child Mnemonic","description":"Generate a BIP85 child mnemonic from a master mnemonic.","operationId":"generate_bip85_child_mnemonic_generate_bip85_child_mnemonic_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/BIP85Request"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/BIP85Response"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"AddressDetails":{"properties":{"derivation_path":{"type":"string","title":"Derivation Path"},"address":{"type":"string","title":"Address"},"public_key":{"type":"string","title":"Public Key"},"private_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Private Key"}},"type":"object","required":["derivation_path","address","public_key"],"title":"AddressDetails"},"AddressListResponse":{"properties":{"addresses":{"items":{"$ref":"#/components/schemas/AddressDetails"},"type":"array","title":"Addresses"}},"type":"object","required":["addresses"],"title":"AddressListResponse"},"AddressRequest":{"properties":{"mnemonic":{"type":"string","title":"Mnemonic"},"passphrase":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Passphrase","default":""},"num_addresses":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Num Addresses","default":1},"include_private_keys":{"type":"boolean","title":"Include Private Keys","default":true}},"type":"object","required":["mnemonic"],"title":"AddressRequest"},"BIP85Request":{"properties":{"mnemonic":{"type":"string","title":"Mnemonic"},"passphrase":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Passphrase","default":""},"app_index":{"type":"integer","title":"App Index","default":39},"word_count":{"type":"integer","title":"Word Count","default":12},"index":{"type":"integer","title":"Index","default":0}},"type":"object","required":["mnemonic"],"title":"BIP85Request"},"BIP85Response":{"properties":{"derivation_path":{"type":"string","title":"Derivation Path"},"child_mnemonic":{"type":"string","title":"Child Mnemonic"}},"type":"object","required":["derivation_path","child_mnemonic"],"title":"BIP85Response"},"BrainWalletRequest":{"properties":{"passphrase":{"type":"string","title":"Passphrase"},"include_private_keys":{"type":"boolean","title":"Include Private Keys","default":true}},"type":"object","required":["passphrase"],"title":"BrainWalletRequest"},"BrainWalletResponse":{"properties":{"bitcoin_address":{"type":"string","title":"Bitcoin Address"},"public_key":{"type":"string","title":"Public Key"},"wif_private_key":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Wif Private Key"}},"type":"object","required":["bitcoin_address","public_key"],"title":"BrainWalletResponse"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MnemonicResponse":{"properties":{"BIP39Mnemonic":{"type":"string","title":"Bip39Mnemonic"},"BIP39Seed":{"type":"string","title":"Bip39Seed"}},"type":"object","required":["BIP39Mnemonic","BIP39Seed"],"title":"MnemonicResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}